Rename components/utils/ => misc/
To keep it distinct from scripts/utils

Hopefully it doesn't break too much on merge :)
