+++ app/scripts/components/boilerplate/language-selector.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import T from '../utils/t';
+import T from '../misc/t';
 
 require('stylesheets/boilerplate/language-selector');
