+++ app/scripts/components/dashboard/polygons.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import {Link} from 'react-router';
-import T from '../utils/t';
+import T from '../misc/t';
 import ChartsContainer from './charts-container';
