+++ app/scripts/components/dashboard/waterpoint.jsx
@@ -1,5 +1,5 @@
 import React,  { PropTypes } from 'react';
-import T from '../utils/t';
+import T from '../misc/t';
 import WaterPoints from './waterpoints';
 
 const WaterPoint = React.createClass({
